refactor: consolidate tag/tags commands into single tags command
Merge the standalone `tag -d` command into the existing `tags` command
as a `-d`/`--delete` option, eliminating the confusing dual command.
